Apply new changes on the original file
